Reduce structure sizing (#58)
* wrap tag in Arc for reduced size

* reduce size of Argument by boxing Hold

* keep argbuilder size but work within boxes to reduce moving big structure around

* Block now just takes a compiler reference

* ArgBuilder takes a compiler reference

* use wildcard pattern matching

* work with a boxed compiler when passing the large structure around

+++ ogma/src/lang/defs.rs
@@ -196,7 +196,7 @@ the row is assumed to be populated with strings, if not an error occurs",
             let help = item.comment.take();
             process_definition(
                 item.code(),
-                Location::File(Arc::clone(&file), item.line as usize),
+                Location::File(Arc::clone(&file), item.line),
                 help,
                 self,
             )?;
